allow float for rate

+++ chatbridge/impl/cqhttp/config.py
@@ -19,5 +19,5 @@ class CqHttpConfig(ClientConfig):
 	qq_whitelist: bool = False
 	mc_whitelist: bool = False
 	limiter: bool = False
-	limiter_rate: int = 2
+	limiter_rate: float = 2
 	limiter_capacity: int = 5
